restructure microservice
