+++ services/file-upload/internal/clients/clients.go
@@ -2,7 +2,7 @@
  * Stuff for configuting and creating S3 clients
  */
 
-package main
+package clients
 
 import (
 	"context"
