+++ services/file-upload/internal/config/storage.go
@@ -1,0 +1,1 @@
+package config
