+++ services/file-upload/internal/storage/minio.go
@@ -1,0 +1,1 @@
+package storage
