+++ services/file-upload/internal/storage/s3.go
@@ -1,0 +1,1 @@
+package storage
