+++ services/file-upload/justfile
@@ -10,7 +10,7 @@ vet:
   go vet ./...
 
 build: vet
-  go build -o {{bindir}}{{pname}} .
+  go build -o {{bindir}}{{pname}} cmd/internal
 
 serve: build
   ./{{bindir}}{{pname}}
